Use actual package data in new demo implementation
AP now has different modes 'AccessPointMenuState'
that will decide what to do when the drum is hit.
This demo implementation broadcasts a simple
OELKAST_LIGHT_SIMPLE_HUE package to the nodes
when the drum is hit. A single RGB Led is used for
testing on both AP and nodes, and the hue is decided
by reading a potentiometer on AP. AP menu state is
changed by pressing button 0.